+++ daml-lf/archive/BUILD.bazel
@@ -111,7 +111,6 @@ da_scala_library(
         ":daml_lf_dev_archive_proto_java",
         "//daml-lf/data",
         "//daml-lf/language",
-        "//libs-scala/logging-entries",
         "@maven//:com_google_protobuf_protobuf_java",
     ],
 )
