+++ daml-lf/data/src/main/scala/com/digitalasset/daml/lf/data/logging/LoggingConfiguration.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.lf.data
+package com.daml.lf.data.logging
 
 case class LoggingConfiguration(
     maxPartyNameLength: Option[Int]
