+++ daml-lf/parser/BUILD.bazel
@@ -49,6 +49,5 @@ da_scala_test(
         ":parser",
         "//daml-lf/data",
         "//daml-lf/language",
-        "//libs-scala/logging-entries",
     ],
 )
