+++ daml-lf/repl/BUILD.bazel
@@ -32,7 +32,6 @@ da_scala_binary(
         "//daml-lf/scenario-interpreter",
         "//daml-lf/transaction",
         "//daml-lf/validation",
-        "//libs-scala/logging-entries",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:org_jline_jline",
     ],
