+++ daml-lf/scenario-interpreter/BUILD.bazel
@@ -31,7 +31,6 @@ da_scala_library(
         "//daml-lf/interpreter",
         "//daml-lf/language",
         "//daml-lf/transaction",
-        "//libs-scala/logging-entries",
         "//libs-scala/nameof",
     ],
 )
