+++ daml-script/test/BUILD.bazel
@@ -146,7 +146,6 @@ da_scala_library(
         "//ledger/sandbox-classic:sandbox-classic-scala-tests-lib",
         "//ledger/sandbox-common",
         "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
-        "//libs-scala/logging-entries",
         "//libs-scala/ports",
         "//libs-scala/resources",
         "@maven//:com_auth0_java_jwt",
