+++ language-support/java/codegen/BUILD.bazel
@@ -100,7 +100,6 @@ da_scala_test(
         "//daml-lf/interface",
         "//language-support/codegen-common",
         "//language-support/java/bindings:bindings-java",
-        "//libs-scala/logging-entries",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:com_squareup_javapoet",
