+++ ledger/participant-integration-api/src/main/scala/platform/indexer/ExecuteUpdate.scala
@@ -312,7 +312,7 @@ trait ExecuteUpdate {
     }
 
   private object Logging {
-    import com.daml.lf.data.Ref.`Party to LoggingKey and LoggingValue`
+    import com.daml.lf.data.logging._
 
     def submissionId(id: SubmissionId): LoggingEntry =
       "submissionId" -> id
